requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
ollama==0.3.3
